test: mock marketMenu separately
